ai/src/training/train_model.py
#!/usr/bin/env python3
"""
Model Training Script for OKR Project

This script handles the training of machine learning models.
"""

import logging
import argparse
from pathlib import Path
from typing import Dict, Any

import torch
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class ModelTrainer:
    """Main class for model training operations."""
    
    def __init__(self, config: Dict[str, Any]):
        """Initialize the model trainer.
        
        Args:
            config: Configuration dictionary with training parameters
        """
        self.config = config
        self.model = None
        self.train_data = None
        self.val_data = None
        
    def load_data(self, data_path: str) -> None:
        """Load training data from file.
        
        Args:
            data_path: Path to the training data
        """
        logger.info(f"Loading data from {data_path}")
        # Implement data loading logic here
        pass
    
    def prepare_data(self) -> None:
        """Prepare data for training (preprocessing, splitting, etc.)."""
        logger.info("Preparing data for training")
        # Implement data preprocessing logic here
        pass
    
    def build_model(self) -> None:
        """Build the machine learning model."""
        logger.info("Building model architecture")
        # Implement model building logic here
        pass
    
    def train(self) -> None:
        """Train the model."""
        logger.info("Starting model training")
        # Implement training logic here
        pass
    
    def evaluate(self) -> Dict[str, float]:
        """Evaluate the trained model.
        
        Returns:
            Dictionary containing evaluation metrics
        """
        logger.info("Evaluating model performance")
        # Implement evaluation logic here
        return {"accuracy": 0.0, "loss": 0.0}
    
    def save_model(self, output_path: str) -> None:
        """Save the trained model.
        
        Args:
            output_path: Path to save the model
        """
        logger.info(f"Saving model to {output_path}")
        # Implement model saving logic here
        pass


def main():
    """Main training function."""
    parser = argparse.ArgumentParser(description="Train ML model for OKR project")
    parser.add_argument("--config", type=str, required=True,
                       help="Path to training configuration file")
    parser.add_argument("--data", type=str, required=True,
                       help="Path to training data")
    parser.add_argument("--output", type=str, required=True,
                       help="Path to save trained model")
    
    args = parser.parse_args()
    
    # Load configuration
    config = {
        "batch_size": 32,
        "learning_rate": 0.001,
        "epochs": 100,
        "model_type": "transformer"
    }
    
    # Initialize trainer
    trainer = ModelTrainer(config)
    
    # Training pipeline
    trainer.load_data(args.data)
    trainer.prepare_data()
    trainer.build_model()
    trainer.train()
    
    # Evaluate and save
    metrics = trainer.evaluate()
    logger.info(f"Training completed. Metrics: {metrics}")
    trainer.save_model(args.output)


if __name__ == "__main__":
    main()