requirements.txt
# Core Python packages
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# AI/ML packages
torch==2.0.1
tensorflow==2.13.0
transformers==4.31.0
huggingface-hub==0.16.4
accelerate==0.21.0

# Data processing
apache-airflow==2.6.3
kafka-python==2.0.2
confluent-kafka==2.1.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pymongo==4.4.1

# API and web frameworks
fastapi==0.101.1
uvicorn==0.23.2
flask==2.3.2
requests==2.31.0
httpx==0.24.1

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Jupyter and notebook tools
jupyter==1.0.0
jupyterlab==4.0.5
ipykernel==6.24.0
notebook==7.0.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
unittest-xml-reporting==3.2.0

# Code quality
black==23.7.0
flake8==6.0.0
mypy==1.5.1
isort==5.12.0
pre-commit==3.3.3

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.1.1
click==8.1.6
tqdm==4.65.0

# Logging and monitoring
loguru==0.7.0
prometheus-client==0.17.1

# Deployment
gunicorn==21.2.0
docker==6.1.3

# Development tools
ipdb==0.13.13
pdbpp==0.10.3